+++ package.json
@@ -4,7 +4,8 @@
   "private": true,
   "main": "index.js",
   "scripts": {
-    "start": "node ./index"
+    "start": "node ./index",
+    "test": "jest --runInBand"
   },
   "dependencies": {
     "body-parser": "~1.19.0",
@@ -14,5 +15,10 @@
     "express": "~4.17.1",
     "morgan": "~1.9.1",
     "pg": "^8.6.0"
+  },
+  "devDependencies": {
+    "jest": "^26.6.3",
+    "superagent": "^6.1.0",
+    "supertest": "^6.1.3"
   }
 }
